all basic functionality completed

+++ app/src/main/java/pzy64/xnotes/ui/screens/trash/TrashFragment.kt
@@ -1,4 +1,4 @@
-package pzy64.xnotes.ui.screens.main
+package pzy64.xnotes.ui.screens.trash
 
 import android.content.Context
 import android.os.Bundle
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.fragment.findNavController
@@ -22,12 +23,12 @@ import pzy64.xnotes.ui.LastItemSpace
 import pzy64.xnotes.ui.baseclasses.Pz64Fragment
 import pzy64.xnotes.ui.screens.MenuType
 import pzy64.xnotes.ui.screens.Pz64ViewModel
+import pzy64.xnotes.ui.screens.main.NotesAdapter
 
-
-class MainFragment : Pz64Fragment() {
+class TrashFragment : Pz64Fragment() {
 
     companion object {
-        fun newInstance() = MainFragment()
+        fun newInstance() = TrashFragment()
     }
 
     private lateinit var viewModel: Pz64ViewModel
@@ -35,8 +36,7 @@ class MainFragment : Pz64Fragment() {
     private lateinit var factory: Pz64ViewModel.Factory
 
     private val placeHolderLoading = "LOADING.."
-    private val placeholderAddNote = "ADD\nNOTE"
-
+    private val placeholderEmpty = "TRASH\nEMPTY"
     private lateinit var noteAdapter: NotesAdapter
 
     private val noteClicked: (Note) -> Unit = { note ->
@@ -48,7 +48,7 @@ class MainFragment : Pz64Fragment() {
 
     private val selectedCallback: (Boolean) -> Unit = { selected ->
         if (selected) {
-            viewModel.menuType.value = MenuType.TypeDelete
+            viewModel.menuType.value = MenuType.TypeDeletePermanent
         } else {
             viewModel.menuType.value = MenuType.TypeNone
         }
@@ -68,7 +68,7 @@ class MainFragment : Pz64Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.main_fragment, container, false)
+        return inflater.inflate(R.layout.trash_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
@@ -85,14 +85,6 @@ class MainFragment : Pz64Fragment() {
         placeHolder.text = placeHolderLoading
         placeHolder.show()
 
-        placeHolder.setOnClickListener {
-            if (placeHolder.text == placeholderAddNote) {
-                viewModel.editMode.value = false
-                viewModel.currentNote.value = null
-                findNavController().navigate(R.id.destinationCreateNote)
-            }
-        }
-
         val layoutManager = StaggeredGridLayoutManager(2, RecyclerView.VERTICAL)
         notesRecyclerView.layoutManager = layoutManager
         val space = resources.getDimensionPixelSize(R.dimen.actionBarSize)
@@ -100,14 +92,14 @@ class MainFragment : Pz64Fragment() {
         noteAdapter = NotesAdapter(noteClicked, selectedCallback)
 
         launch {
-            viewModel.getNotes().observe(this@MainFragment, Observer {
+            viewModel.getTrash().observe(this@TrashFragment, Observer {
                 it?.let { notes ->
                     if (notes.isNotEmpty()) {
                         noteAdapter.data = notes.toMutableList()
                         notesRecyclerView.adapter = noteAdapter
                         placeHolder.hide()
                     } else {
-                        placeHolder.text = placeholderAddNote
+                        placeHolder.text = placeholderEmpty
                         placeHolder.show()
                     }
                 }
@@ -117,24 +109,49 @@ class MainFragment : Pz64Fragment() {
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
 
-        when(item.itemId) {
-            R.id.actionDelete -> {
-               noteAdapter.deleteSelectedNotes()
+        when (item.itemId) {
+            R.id.actionDeleteForever -> {
+                if (attached) {
+                    AlertDialog.Builder(context!!)
+                        .setTitle("Confirm Deletion")
+                        .setMessage("Deleted note can't be retrieved later.")
+                        .setPositiveButton(android.R.string.ok) { dialog, which ->
 
+                            launch {
+                                val notes = viewModel.notes.filter {
+                                    it.isSelected
+                                }
+                                viewModel.deleteNote(notes)
+                                viewModel.menuType.value = MenuType.TypeNone
+                                noteAdapter.deleteSelectedNotes()
+                                if (viewModel.notes.isEmpty()) {
+                                    placeHolder.text = placeholderEmpty
+                                    placeHolder.show()
+                                }
+                            }
+                        }.setNegativeButton(android.R.string.cancel) { dialog, which ->
+                            dialog.dismiss()
+                        }
+                        .setOnDismissListener {
+                            noteAdapter.deSelectAll()
+                            viewModel.menuType.value = MenuType.TypeNone
+                        }.show()
+                }
 
+            }
+            R.id.actionRestore -> {
+                noteAdapter.deleteSelectedNotes()
                 launch {
                     val notes = viewModel.notes.filter {
                         it.isSelected
                     }
-                    viewModel.moveToTrash(notes)
-                    viewModel.menuType.value = MenuType.TypeNone
-
+                    viewModel.restoreFromTrash(notes)
                     if (viewModel.notes.isEmpty()) {
-                        placeHolder.text = placeholderAddNote
+                        placeHolder.text = placeholderEmpty
                         placeHolder.show()
                     }
+                    viewModel.menuType.value = MenuType.TypeNone
                 }
-
             }
         }
